+++ CHANGELOG.md
@@ -142,6 +142,7 @@
 - ``Broken Pipe`` signal message after evaluation of last-pipe
 - infix keyword completion
 - PATH handling of sudo completion
+- length error of brace expansion
 
 ## [0.26.0] - 2022-06-30